Adding asynchronous support
This is a major change, adds a dependency to the after package for async
support, all hooks are now expected to call the next callback function
when they are ready to exit.
All of the hooks will run even if a hook doesn't call the next function
and it's unnecessary to implement any timeout logic since most, if not
all, process managers will send a SIGKILL if the SIGTERM takes to long.

Note that if someone explicitly terminates the process with SIGKILL or
process.exit() the asynchronous code will not finish running its code.